requirements.txt
numpy
pandas
xbgoost
seaborn
lightgbm
matplotlib
setuptools
scikit-learn
mysql-connector-python